// frontend/src/shared/types/store.ts
export interface SocketState {
  socket: Socket | null;
  onlineUsers: string[];
  showOnlineStatus: boolean;
  connectSocket: () => void;
  loadShowOnlineStatus: () => Promise<void>;
  updateShowOnlineStatus: (value: boolean) => Promise<void>;
  emitShowOnlineStatus: (value: boolean) => void;
  emitActiveConversation: (conversationId: string | null) => void;
  disconnectSocket: () => void;
}